+++ backend/canisters/airdrop_bot/api/src/updates/set_airdrop.rs
@@ -1,3 +1,4 @@
+use crate::{AirdropAlgorithm, AirdropConfig};
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{ChannelId, CommunityId, TimestampMillis};
@@ -7,10 +8,7 @@ pub struct Args {
     pub community_id: CommunityId,
     pub channel_id: ChannelId,
     pub start: TimestampMillis,
-    pub main_chat_fund: u128,
-    pub main_chit_band: u32,
-    pub lottery_prizes: Vec<u128>,
-    pub lottery_chit_band: u32,
+    pub algorithm: AirdropAlgorithm,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -19,4 +17,15 @@ pub enum Response {
     ChannelUsed,
     InThePast,
     ClashesWithPrevious,
+}
+
+impl From<Args> for AirdropConfig {
+    fn from(value: Args) -> Self {
+        AirdropConfig {
+            community_id: value.community_id,
+            channel_id: value.channel_id,
+            start: value.start,
+            algorithm: value.algorithm,
+        }
+    }
 }